Refactor imports for @hedvig-ui/SearchableDropdown
