+++ src/components/claims/claim-details/components/ClaimType.tsx
@@ -6,13 +6,18 @@ import {
   useSetClaimTypeMutation,
 } from 'api/generated/graphql'
 
-import { Button, CardContent, DateTimePicker, Input } from '@hedvig-ui'
+import {
+  Button,
+  CardContent,
+  DateTimePicker,
+  Input,
+  SearchableDropdown,
+} from '@hedvig-ui'
 import {
   PaperTitle,
   PaperTitleBadgeProps,
 } from 'components/claims/claim-details/components/claim-items/PaperTitle'
 import { format, parseISO } from 'date-fns'
-import { SearchableDropdown } from 'hedvig-ui/searchable-dropdown'
 import { Spacing } from 'hedvig-ui/spacing'
 import { Label } from 'hedvig-ui/typography'
 import React, { useState } from 'react'
