+++ src/features/tools/campaign-codes/forms/FreeMonthsForm.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import { Button, DateTimePicker, Input } from '@hedvig-ui'
+import { Button, DateTimePicker, Input, SearchableDropdown } from '@hedvig-ui'
 import { AssignVoucherFreeMonths, Scalars } from 'api/generated/graphql'
 import { PartnerDropdown } from 'features/tools/campaign-codes/forms/PartnerDropdown'
 import { getCodeTypeOptions } from 'features/tools/campaign-codes/utils'
@@ -7,7 +7,6 @@ import {
   addPartnerFreeMonthsCodeOptions,
   useAddPartnerFreeMonthsCode,
 } from 'graphql/use-add-partner-free-months-code'
-import { SearchableDropdown } from 'hedvig-ui/searchable-dropdown'
 import { Spacing } from 'hedvig-ui/spacing'
 import { Label } from 'hedvig-ui/typography'
 import React from 'react'
