+++ src/features/tools/employees/components/CreateEmployee.tsx
@@ -1,12 +1,11 @@
 import styled from '@emotion/styled'
-import { Button, ButtonsGroup } from '@hedvig-ui'
+import { Button, ButtonsGroup, SearchableDropdown } from '@hedvig-ui'
 import {
   EmployeesDocument,
   useAvailableEmployeeRolesQuery,
   useCreateEmployeeMutation,
 } from 'api/generated/graphql'
 import { Row } from 'features/tools/employees'
-import { SearchableDropdown } from 'hedvig-ui/searchable-dropdown'
 import React, { useState } from 'react'
 import { toast } from 'react-hot-toast'
 import { Input } from 'semantic-ui-react'
