+++ src/features/tools/employees/components/EmployeeFilter.tsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled'
+import { SearchableDropdown } from '@hedvig-ui'
 import { useAvailableEmployeeRolesQuery } from 'api/generated/graphql'
 import { Row } from 'features/tools/employees'
 import { CreateEmployee } from 'features/tools/employees/components/CreateEmployee'
-import { SearchableDropdown } from 'hedvig-ui/searchable-dropdown'
 import React from 'react'
 import { Input } from 'semantic-ui-react'
